fix issues with UEFI firmware image

+++ kernel/src/arch/x86_64/memory/paging.cpp
@@ -492,7 +492,7 @@ void PageMap::create_new(PageMap* map) {
     // half of the address space by copying upper-level entries.
     if (kernel_initialized) {
         uintptr_t kmap       = get_kernel_map()->phys_root_addr;
-        uintptr_t* kmap_virt = reinterpret_cast<uintptr_t*>(to_higher_half(kmap));
+        uint64_t* kmap_virt = reinterpret_cast<uint64_t*>(to_higher_half(kmap));
 
         for (int i = 256; i < 512; ++i) {
             root_virt[i] = kmap_virt[i];
